Items Control should have a Transp. Background so it can be used as a Drop Target
